Craete UI.java
Consolidate UI functions in UI.java
