--- src/main/java/Farewell.java
@@ -1,7 +1,0 @@
-public class Farewell {
-  public static String farewell() {
-    return """
-           Later young lass. Hope ta see ya around these here parts!
-           """;
-  }
-}
