--- src/main/java/errors/ErrorHandler.java
@@ -1,9 +1,0 @@
-package errors;
-
-public class ErrorHandler {
-  public static String getErrorMessage(BoopError error) {
-    return """
-        Arr me bad! %s
-        """.formatted(error.getMessage());
-  }
-}
